5.5.317

+++ OKX.Api/Account/Responses/OkxAccountLeverage.cs
@@ -27,5 +27,5 @@ public record OkxAccountLeverage
     /// Leverage
     /// </summary>
     [JsonProperty("lever")]
-    public decimal Leverage { get; set; }
+    public int Leverage { get; set; }
 }
