+++ OKX.Api/CopyTrading/Responses/OkxCopyTradingLeadTraderPosition.cs
@@ -33,7 +33,7 @@ public record OkxCopyTradingLeadTraderPosition
     /// Leverage
     /// </summary>
     [JsonProperty("lever")]
-    public decimal Leverage { get; set; }
+    public int Leverage { get; set; }
 
     /// <summary>
     /// Average open price
