+++ OKX.Api/Grid/Clients/OkxGridRestClient.cs
@@ -98,7 +98,7 @@ public class OkxGridRestClient(OkxRestApiClient root) : OkxBaseRestClient(root)
         // Contract Grid Order
         decimal? size = null,
         OkxGridContractDirection? contractGridDirection = null,
-        decimal? leverage = null,
+        int? leverage = null,
         bool? basePosition = null,
         decimal? takeProfitRatio = null,
         decimal? stopLossRatio = null,
@@ -561,7 +561,7 @@ public class OkxGridRestClient(OkxRestApiClient root) : OkxBaseRestClient(root)
         decimal gridNumber,
         OkxGridRunType? gridRunType,
         OkxGridContractDirection? direction = null,
-        decimal? leverage = null,
+        int? leverage = null,
         bool? basePosition = null,
         OkxGridInvestmentType? investmentType = null,
         OkxGridTriggerStrategy? triggerStrategy = null,
