+++ OKX.Api/Grid/Responses/OkxGridComputedMarginBalance.cs
@@ -9,7 +9,7 @@ public record OkxGridComputedMarginBalance
     /// Leverage
     /// </summary>
     [JsonProperty("lever")]
-    public decimal Leverage { get; set; }
+    public int Leverage { get; set; }
 
     /// <summary>
     /// Maximum quantity
